Changes RoomOccupancy to Availability This changes the RoomOccupancy service to an Availability service which only returns room type ids that have availability for a given date.
