+++ src/Availability.Api/App_Start/WebApiConfig.cs
@@ -1,0 +1,21 @@
+﻿using System.Web.Http;
+
+namespace Availability.Api
+{
+	public static class WebApiConfig
+	{
+		public static void Register(HttpConfiguration config)
+		{
+			// Web API configuration and services
+
+			// Web API routes
+			config.MapHttpAttributeRoutes();
+
+			config.Routes.MapHttpRoute(
+				"DefaultApi",
+				"api/{controller}/{id}",
+				new {id = RouteParameter.Optional}
+			);
+		}
+	}
+}
