+++ src/Availability.Api/Global.asax
@@ -1,0 +1,1 @@
+﻿<%@ Application Codebehind="Global.asax.cs" Inherits="Availability.Api.WebApiApplication" Language="C#" %>
