+++ src/Availability.Api/Global.asax.cs
@@ -1,0 +1,13 @@
+﻿using System.Web;
+using System.Web.Http;
+
+namespace Availability.Api
+{
+	public class WebApiApplication : HttpApplication
+	{
+		protected void Application_Start()
+		{
+			GlobalConfiguration.Configure(WebApiConfig.Register);
+		}
+	}
+}
