+++ src/Availability.Api/Models/DateRange.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace Availability.Api.Models
+{
+	public class DateRange
+	{
+		public DateTime StartDate { get; set; }
+		public DateTime EndDate { get; set; }
+	}
+}
