+++ src/Availability.Api/Models/SearchCriteria.cs
@@ -1,0 +1,10 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Availability.Api.Models
+{
+	public class SearchCriteria
+	{
+		[Required]
+		public DateRange Dates { get; set; }
+	}
+}
