+++ src/Availability.Api/Web.Release.config
