+++ src/Availability.Api/favicon.ico
