+++ src/Availability.Api/packages.config
