+++ src/Availability/Messages/Events/RoomTypeHoldSubmittedEvent.cs
@@ -1,0 +1,7 @@
+namespace Availability.Messages.Events
+{
+	public class RoomTypeHoldSubmittedEvent
+	{
+		public int RoomTypeId { get; set; }
+	}
+}
