+++ src/Availability/nservicebus.host.exe.config
