--- src/RoomOccupancy.Api/Global.asax
@@ -1,1 +1,0 @@
-﻿<%@ Application Codebehind="Global.asax.cs" Inherits="RoomOccupancy.Api.WebApiApplication" Language="C#" %>
