--- src/RoomOccupancy.Api/Global.asax.cs
@@ -1,13 +1,0 @@
-﻿using System.Web;
-using System.Web.Http;
-
-namespace RoomOccupancy.Api
-{
-	public class WebApiApplication : HttpApplication
-	{
-		protected void Application_Start()
-		{
-			GlobalConfiguration.Configure(WebApiConfig.Register);
-		}
-	}
-}
