--- src/RoomOccupancy.Api/Models/DateRange.cs
@@ -1,10 +1,0 @@
-﻿using System;
-
-namespace RoomOccupancy.Api.Models
-{
-	public class DateRange
-	{
-		public DateTime StartDate { get; set; }
-		public DateTime EndDate { get; set; }
-	}
-}
