--- src/RoomOccupancy.Api/Models/SearchCriteria.cs
@@ -1,7 +1,0 @@
-﻿namespace RoomOccupancy.Api.Models
-{
-	public class SearchCriteria
-	{
-		public DateRange Dates { get; set; }
-	}
-}
