--- src/RoomOccupancy.Data/Context/SqLiteConfig.cs
@@ -1,17 +1,0 @@
-﻿using System.Data.Entity;
-using System.Data.Entity.Core.Common;
-using System.Data.SQLite;
-using System.Data.SQLite.EF6;
-
-namespace Divergent.Finance.Data.Context
-{
-    internal class SqLiteConfig : DbConfiguration
-    {
-        public SqLiteConfig()
-        {
-            SetProviderFactory("System.Data.SQLite", SQLiteFactory.Instance);
-            SetProviderFactory("System.Data.SQLite.EF6", SQLiteProviderFactory.Instance);
-            SetProviderServices("System.Data.SQLite", (DbProviderServices)SQLiteProviderFactory.Instance.GetService(typeof(DbProviderServices)));
-        }
-    }
-}
