--- src/RoomOccupancy.Data/Models/RoomType.cs
@@ -1,9 +1,0 @@
-﻿namespace RoomOccupancy.Data.Models
-{
-	public class RoomType
-	{
-		public int Id{ get; set; }
-		public string Name { get; set; }
-		public int NumAvailable { get; set; }
-	}
-}
