--- src/RoomOccupancy/Config/EndpointConfig.cs
@@ -1,64 +1,0 @@
-﻿using System;
-using System.IO;
-using System.Linq;
-using NServiceBus;
-using NServiceBus.Logging;
-using RoomOccupancy.Data.Context;
-using ILog = Common.Logging.ILog;
-using LogManager = Common.Logging.LogManager;
-
-namespace RoomOccupancy.Config
-{
-	[EndpointName("HotelManagementSystem.RoomOccupancy")]
-	public class EndpointConfig : IConfigureThisEndpoint, AsA_Server
-	{
-		private static readonly ILog Log = LogManager.GetLogger<EndpointConfig>();
-
-		public EndpointConfig()
-		{
-			NServiceBus.Logging.LogManager.Use<DefaultFactory>();
-
-			if (Environment.UserInteractive)
-				Console.Title = "Room Occupancy";
-
-			InitializeDatbase();
-		}
-
-		private void InitializeDatbase()
-		{
-			Log.Debug("Initializing database");
-
-			var context = new RoomOccupancyContext();
-			var roomTypes = context.RoomTypes.ToList();
-
-			Log.DebugFormat("Database initialized, first room type is {0}", roomTypes.First().Id);
-		}
-
-		public void Customize(EndpointConfiguration endpointConfiguration)
-		{
-			Log.Info("Customizing endpoint...");
-
-			var container = ContainerSetup.Create();
-
-			var licensePath = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, "License.xml");
-			endpointConfiguration.LicensePath(licensePath);
-			endpointConfiguration.UseSerialization<JsonSerializer>();
-			endpointConfiguration.Recoverability().Delayed(c => c.NumberOfRetries(0));
-			endpointConfiguration.UseContainer<AutofacBuilder>(c => c.ExistingLifetimeScope(container));
-			endpointConfiguration.UseTransport<MsmqTransport>()
-				.ConnectionString("deadLetter=false;journal=false");
-			//endpointConfiguration.UsePersistence<NHibernatePersistence>()
-			//    .ConnectionString(ConfigurationManager.ConnectionStrings["HotelManagementSystem.RoomOccupancy"].ToString());
-			endpointConfiguration.UsePersistence<InMemoryPersistence>();
-
-			endpointConfiguration.SendFailedMessagesTo("error");
-			endpointConfiguration.AuditProcessedMessagesTo("audit");
-
-			var conventions = endpointConfiguration.Conventions();
-			conventions.DefiningCommandsAs(t => t.Namespace != null && t.Namespace.StartsWith("HotelManagementSystem") && t.Namespace.EndsWith("Commands") && t.Name.EndsWith("Command"));
-			conventions.DefiningEventsAs(t => t.Namespace != null && t.Namespace.StartsWith("HotelManagementSystem") && t.Namespace.EndsWith("Events") && t.Name.EndsWith("Event"));
-
-			endpointConfiguration.EnableInstallers();
-		}
-	}
-}
