--- src/RoomOccupancy/Messages/Events/RoomTypeHoldSubmittedEvent.cs
@@ -1,7 +1,0 @@
-namespace RoomOccupancy.Messages.Events
-{
-	public class RoomTypeHoldSubmittedEvent
-	{
-		public int RoomTypeId { get; set; }
-	}
-}
